backend/fast-api-audio/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
librosa==0.10.1
praat-parselmouth==0.4.3
numpy==1.26.2
matplotlib==3.8.2